Move monoids module to a separate file
